move _num_executions counter to execute() method for _device and _qubit_device, move corresponding unit tests to appropriate locations
